appledoc beginnings.
I've begun to convert the documentation for TwUI over to appledoc. I've
done most of TUIScrollView so far because it was the first one that
threw tons of errors when I initially ran appledoc.

I'm using this to compile the docs:
› appledoc --project-name "TwUI" --project-company "Twitter, Inc."
--company-id "com.twitter" --output "./docs_new" ./lib

I've moved many comment blocks from their .m files to the .h files and
combined multiple descriptions of setters and getters for properties
into a single description.

I'm using the UIKit documentation as a guide for organization. At the
moment I'm using mostly the docs that are already there, but I have
written some myself for completeness (but they do not come from the
UIKit docs of their counterparts).